Decided to check for presence of keypresses using global hook.
So the decision has to be make whether we ask os that the key is pressed
or we just use our own hookup.

The latter is much much easier, but not sure if its reliable.
We shall see.

However since the implementation of hotkeys will not use OS abilities, i
guess this is at least consistent.